ZH: 支持单人突击挑战模式
