+++ module/solo_raid/challenge.py
@@ -3,28 +3,37 @@ from module.logger import logger
 from module.ocr.ocr import Digit
 from module.simulation_room.assets import AUTO_BURST, AUTO_SHOOT, END_FIGHTING
 from module.solo_raid.assets import *
+from module.ui.assets import MAIN_CHECK
 from module.ui.ui import UI
 
 
 class SoloRaidChallenge(UI):
     @property
+    def free_remain(self) -> int:
+        model_type = self.config.Optimization_OcrModelType
+        FREE_REMAIN = Digit(
+            [FREE_OPPORTUNITY_CHECK.area],
+            name='FREE_REMAIN',
+            model_type=model_type,
+            lang='ch',
+        )
+        return int(FREE_REMAIN.ocr(self.device.image)['text'])
+
+    @property
+    def free_opportunity_remain(self) -> bool:
+        # result = self.appear(FREE_OPPORTUNITY_CHECK, offset=10, threshold=0.8)
+        if self.free_remain:
+            logger.info(f'[Free opportunities remain] {self.free_remain}')
+        return self.free_remain
+    
+    @property
     def challenge_damage_is_zero(self) -> bool:
         """
         判断挑战模式当前伤害是否为0
         """
-        try:
-            model_type = self.config.Optimization_OcrModelType
-            CURRENT_DAMAGE = Digit(
-                [CURRENT_DAMAGE_AREA.area],  # 请确保在assets.py中定义了 CURRENT_DAMAGE_AREA
-                name='CURRENT_DAMAGE',
-                model_type=model_type,
-                lang='ch',
-            )
-            damage = int(CURRENT_DAMAGE.ocr(self.device.image)['text'])
-            logger.info(f'[Current Challenge Damage] {damage}')
-            return damage == 0
-        except Exception as e:
-            logger.warning(f'Failed to read damage or damage is empty: {e}')
+        if self.appear(CHALLENGE_DAMAGE_ZERO, offset=10):
+            return True
+        else:
             return False
 
     def ensure_into_challenge(self, skip_first_screenshot=True):
@@ -38,38 +47,37 @@ class SoloRaidChallenge(UI):
             else:
                 self.device.screenshot()
 
-            # 检查当前是否在第七关界面，不是则直接退出挑战逻辑
-            if self.appear(SOLO_RAID_CHECK, offset=10) and not self.appear(STAGE_SEVEN, offset=(30, 30)):
-                logger.info('Current stage is not Stage 7. Skip challenge mode.')
-                return
-
-            # 如果在第七关，点击 STAGE_CHALLENGE_SWITCH 切换到挑战模式
             if (
                 click_timer.reached()
-                and self.appear(SOLO_RAID_CHECK, offset=10)
-                and self.appear(STAGE_SEVEN, offset=(30, 30))
-                and self.appear_then_click(STAGE_CHALLENGE_SWITCH, offset=10, interval=2)
+                and self.appear(MAIN_CHECK, offset=10)
+                and self.appear_then_click(SOLO_RAID, offset=10, interval=3, static=False)
             ):
-                click_timer.reset()
+                logger.info('Enter solo raid')
                 continue
 
-            # 确认已经进入挑战模式页面
-            if self.appear(STAGE_CHALLENGE, offset=(30, 30)):
-                break
+            # 检查挑战
+            if click_timer.reached() and self.appear(SOLO_RAID_CHECK, offset=(10, 10)):
+                if self.appear(STAGE_CHALLENGE, offset=(30, 30)):
+                    # 挑战界面
+                    break
+                else:
+                    # 挑战未开启
+                    logger.warning('Challenge mode not reached')
+                    return
 
-        # 1. 检查伤害是否为0
+        # 检查伤害是否为0
         if not self.challenge_damage_is_zero:
             logger.info('Challenge damage is already recorded. Skip challenge mode.')
             return
 
-        # 2. 检查挑战次数 (free_opportunity_remain 由主类共享)
+        # 检查挑战次数
         if self.free_opportunity_remain:
             self.challenge_raid()
         else:
             logger.warning('There are no free opportunities for challenge mode')
 
     def challenge_raid(self, skip_first_screenshot=True):
-        """挑战模式战斗执行 (无扫荡逻辑)"""
+        """挑战模式战斗执行"""
         logger.hr('Start a challenge raid')
         click_timer = Timer(0.3)
